+++ NEWS.md
@@ -3,6 +3,7 @@
 # zipcodeR 0.1.1.9000 (WIP)
 - Add function for obtaining congressional districts of a given ZIP code.
 - Add function for obtaining ZIP codes contained within a congressional district.
+- Add function for obtaining county and state names for a provided ZIP code
 
 # zipcodeR 0.1.0 (2020-09-08)
 Initial public release, first version accepted by CRAN.